generalize data upload utils to be used outside of the file based use case
